requirements.txt
discord.py>=2.3.0
python-dotenv>=1.0.0
requests>=2.32.3
aiohttp
llama-cpp-python
